Add debian packages build system to wrap all Whale Tag specific software.
Introduce a ceti-tag-set-hostname systemd service that changes hostname
to one derived from the unique device id of the system it runs on.
The following ID is used:
  1) WLAN MAC address
  2) Ethernet MAC address
  3) Serial entry in /proc/cpuinfo
If none of those exists, the hostname is left as is.

The hostname will be used to create folder structure for data capture.
Having each Whale Tag have its own unique id will help implement
the data import pipeline, since each data stream is expected to
be located in its own folder in the cloud for future processing.