Introducing sdcard build system
The build start by constructing a docker environment for the
repeatbale builds. Then raspbian image is downloaded.
The QEMU is lunched from the inside of the Docker container
Then a setup_image.sh is executed inside the image under QEMU
to install all the necessary software. Out comes the sdcard
image for flashing onto the raspberry pi sd card.