Implement mutex-less page flipping.
Now use two pages to acquire data from FPGA through SPI into.
Switch between the pages, once one is filled.
The filled page is saved to disk, while the other page can
be used to keep acquiring data into.

No mutexes used. So in principal this is still not thread-safe.
However we seem to have never had collsions before, even with one page.

This potentially could become a problem, if saving to disk
takes a VERY long time. For example, if storage device needs to be
woken up from low power state. But we can cross that brindge when
we get there.