dos2unix files before building.
This is silly, but the repo was started in Windows CR/LF,
and git does not seem to be able to properly deal with different
line breaks. As such, it breaks the bash scripts that awk the
name of the debian package during build.

So as a workaround, dos2unix everything that matters before building.