WIP: Use FLAC streaming encoder for all audio data.
It builds and runs, but I have no hardware to properly
test it on. We need to test this, an measure CPU consumption.
Additionally, I am sure this can be made more efficient
if the main system had more RAM. Currently I call encoder with
only one set of samples at a time - because of RAM limitations.