WIP: switch to arm64 based raspios instead of armv6l raspbian-lite
This is work in progress commit that switches the hw target from
raspberry pi zero W to raspberry pi zero 2.

The only problem seems to be with the actual C code we produced
for the main CETI app. Linker is currently unhappy.
But I wanted to push this code here to unblock others to make progress.