Pin audioSpiThread to CPU0 using set_affinity_np().
On raspberry pi zero w 2, according to Joseph, FIFO overflows
still occur. The current hypothesis is that cpu misses the deadline
to drain FPGA FIFO during the thread being moved by the scheduler
between the cores. Attempt to remedy this by only allowing the scheduler
to use CPU0 for the audioSpiThread, and keep its priority highest.