Fix ceti-tag-battery-monitor.
Make sure the service starts, and the bash script has correct permissions.
Also make sure the pigpio library initialized properly.
The only problem with this thing is that by default it returns 0.0 for volatges,
which results in immediate shutdown of the system, if the gas gauge is not present.
Consider other default values.