Fix the flac encoder
The issue was in the function signiture of the spi_read() function which
expected char type. Later we converted this into 16bit integers to pass
to the encoder, and that resulted in wrong numerical representation.

In the cl I brought back the original large buffer for the encoder to
use, since we have swap now and are not as memory constricted. We can
rework it again to lower memory footprint, if it will start causing
issues.